feat(rename-r3-stage-b3): codons.py -> fragments.py + shim (#87)
R3 Stage B.3 per the plan in ~/.claude/plans/ethereal-forging-cookie.md
and tracking issue #87.

Files:
  - helix_context/codons.py -> helix_context/fragments.py (rename)
  - helix_context/codons.py  (new shim — re-exports everything)

Class names (Codon / CodonChunker / CodonEncoder / RawStrand) and
the Pydantic field name (Gene.codons / Document.codons) are unchanged
— Stage C may rename the helper class identifiers later.

Verification:
  - import surfaces resolve via Codon (codons), Codon (fragments),
    and helix_context.Codon all to the same class
  - tests/test_codons.py + test_genome.py: 55/55 pass in 0.4s
